refactor: Move Mergekit integration to experimental submodule
Resolves #4395

- Move mergekit_utils.py to trl/experimental/mergekit/
- Move MergeModelCallback to trl/experimental/mergekit/callbacks.py
- Create trl/experimental/mergekit/__init__.py for module exports
- Remove MergeModelCallback from main trl exports (trl/__init__.py, trl/trainer/__init__.py)
- Remove MergeModelCallback from trl/trainer/callbacks.py
- Update test imports to use trl.experimental.mergekit
- Update callbacks documentation with experimental status note

This change consolidates underutilized Mergekit functionality into the
experimental namespace as part of v1.0 preparations. The module may be
deprecated or removed in future versions if usage does not materialize.

To use MergeModelCallback after this change:
```python
from trl.experimental.mergekit import MergeConfig, MergeModelCallback
```
