+++ trl/__init__.py
@@ -110,7 +110,6 @@ _import_structure = {
     ],
     "trainer.callbacks": [
         "BEMACallback",
-        "MergeModelCallback",
         "RichProgressCallback",
         "SyncRefModelCallback",
         "WeaveCallback",
@@ -191,7 +190,6 @@ if TYPE_CHECKING:
     )
     from .trainer.callbacks import (
         BEMACallback,
-        MergeModelCallback,
         RichProgressCallback,
         SyncRefModelCallback,
         WeaveCallback,
