+++ trl/trainer/__init__.py
@@ -23,7 +23,6 @@ _import_structure = {
     "callbacks": [
         "BEMACallback",
         "LogCompletionsCallback",
-        "MergeModelCallback",
         "RichProgressCallback",
         "SyncRefModelCallback",
         "WeaveCallback",
@@ -83,7 +82,6 @@ if TYPE_CHECKING:
     from .callbacks import (
         BEMACallback,
         LogCompletionsCallback,
-        MergeModelCallback,
         RichProgressCallback,
         SyncRefModelCallback,
         WeaveCallback,
